requirements.txt
torch
torchvision
pyyaml
numpy
opencv-python
Pillow
